Add matching abstract class so that new implementations only need to override the tiebreaker strategy
